Kickstart information
Informations that are important and basic should be found fast. Why it is necessary? Sometimes css is not used only in webapps and scripts are not only used in webapps, so they are located in `src/main/scripts` or in `src/main/resources`. See https://maven.apache.org/guides/introduction/introduction-to-the-standard-directory-layout.html